s1: Fixed windows build bugs.

+++ trc/arith.c
@@ -275,5 +275,5 @@ int32_t uint32_to_int32(uint32_t x) {
 
 size_t uint64_to_size(uint64_t x) {
   CHECK(x <= SIZE_MAX);
-  return x;
+  return (size_t)x;
 }
